Add simple memory visualiser for debugging

+++ src/main/java/org/os/core/Memory.java
@@ -3,7 +3,7 @@ package org.os.core;
 public interface Memory {
     Word read(int address);
 
-    void write(int address, int value);
+    void write(int address, long value);
 
     void show();
 }
